requirements.txt
# FastAPI & Server
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# Database
supabase>=2.4.0
qdrant-client>=1.9.0

# AI / LLM
google-generativeai>=0.5.0
langchain>=0.2.0
langchain-google-genai>=1.0.0
langchain-community>=0.2.0

# Utilities
pydantic>=2.7.0
pydantic-settings>=2.2.0
pydantic[email]>=2.7.0
python-dotenv>=1.0.0
httpx>=0.27.0
Pillow>=10.3.0
PyPDF2>=3.0.0
python-jose[cryptography]>=3.3.0
PyJWT>=2.8.0

# Scheduling / Notifications
apscheduler>=3.10.0
resend>=2.0.0


#increption
pycryptodome>=3.19.0